refactor: props 선언 및 스토리북에 더미데이터 수정
