+++ src/stories/organisms/SearchTalkPickListSection.stories.tsx
@@ -1,13 +1,29 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
 import SearchTalkPickListSection from '@/components/organisms/SearchTalkPickListSection/SearchTalkPickListSection';
+import { SampleWhole } from '@/assets';
+import { SearchTalkPickItemProps } from '@/components/atoms/SearchTalkPickItem/SearchTalkPickItem';
+
+const SearchTalkPickItems: SearchTalkPickItemProps[] = Array.from(
+  { length: 30 },
+  (_, index) => ({
+    title: '월클 정국 VS 존잘 차은우',
+    date: '2024.08.26',
+    content:
+      '우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하우하하',
+    imgUrl: SampleWhole,
+  }),
+);
 
 const meta = {
   title: 'organisms/SearchTalkPickListSection',
   component: SearchTalkPickListSection,
 
   tags: ['autodocs'],
-  args: {},
+  args: {
+    searchTalkPickList: SearchTalkPickItems,
+  },
 } satisfies Meta<typeof SearchTalkPickListSection>;
 
 export default meta;
